Add files via upload
Deterministic_model: allows user to run the non-spatial version of the combined behavioural and epidemic model. The user can choose between the four behavioural models outlined in the paper ("strategy vs. population", "strategy vs. alternative", "grower vs. population" and "grower vs. alternative" (which is the main focus of the paper). 

Spatial_stochastic_model: allows users to run the "grower vs. alternative" version of the spatial-stochastic model outlined in the paper.