Add files via upload
CSS participation and proportion of infected fields are dependent on economic and epidemi-
ological parameters. Plots show the equilibrium proportion of controllers in response to variations in
parameters when comparisons are made using the “grower vs. alternative” comparison. (A) Effect
of the rate of horizontal transmission (β) and cost of control (φ). (B) Effect of the rate of horizontal
transmission (β) and probability of receiving infected planting material via trade (p). (C) Effect of
the rate of horizontal transmission (β) and loss due to infection (L). Parameters are as in Table 1.