Add files via upload
Effect of responsiveness (η) on the “grower vs. alternative” model. (A) The flow of growers
between the non-infected controllers (SC ), infected controllers (IC ) and infected non-controllers (IN )
based on their probability of switching strategy (zSC , zIC and zIN respectively) with η = 1. To Get (A) and (C), multiply the respective switching terms and state variables together (i.e. SC * zSC).