Add files via upload
Model dynamics for different means of comparison. Here, SN are susceptible non-
controllers, IN infected non-controllers, SC are susceptible controllers and IC infected controllers.
C represents SC + IC . In (A), growers compare the expected profit of their strategy with that of
the entire population; in (B) they compare the expected profit of their strategy with the expected
profit of the alternative strategy. For (C), growers compare their own outcome with expected profit
of the population and in (D) they compare their outcome with the expected value of the alternative
strategy. Parameters are as in Table 1 in the main text.