Add files via upload
Effect of systematic misestimation in the spatial-stochastic model on (A)  the perceptions of the dispersal scale for the whitefly vector (να) increase and (B) the rate of horizontal transmission increase (νβ ). The mean values were calculated
over 100 runs are the error bars show one standard deviation around the mean. Parameters are as in Table 4, Appendix 2.