Add files via upload
Probability of infection, responsiveness of growers, expected profits, and switching terms
for the “strategy vs. population” model. (A) Probability of infection for the default parameterisation. (B)
The expected profits for controllers, non-controllers and the population for different proportions
of controllers in the population (c). (C)
Switching probabilities for controllers and non-controllers for different c. (D) The responsiveness of growers (η). (E) The expected profits
for controllers, non-controllers and the population when the cost of control φ, = 0.125.  (F) Switching terms for φ = 0.125. Other parameters are as in Table 1.