Add files via upload
Comparison of default behaviour of the “grower vs. alternative” spatial-stochastic and
deterministic models. (A) + (B) Dynamics for spatial model with phi = 0.25 and 0.125 respectively.  (C) and (D) are the same for the deterministic model. The figures show the mean for 100 runs of each model, and the error bars show one standard deviation. Parameters are as in Table 1 (main text) and Table 4 (Appendix 2).