Add files via upload
Possible equilibria for the “grower vs. population” model when p = 1. There can now be a disease-endemic, no-control equilibrium for this dataset, but now that p = 1, there can never be a disease-free equilibrium (as R0 > 1). Parameters are as in Table 1. 2= all disease, no control equilibrium and 4 = disease, control and no-control equilibrium.